+++ src/pipecatcloud/cli/api.py
@@ -7,4 +7,4 @@
 from pipecatcloud.api import _API
 from pipecatcloud.cli.config import config
 
-API = _API(config.get("token"))
+API = _API(config.get("token"), is_cli=True)